Fix a crash when entering fullscreen mode with MediaRemote
A crash occurs sometimes when entering fullscreen mode with MediaRemote.
It is caused when the parent of the dialog is no longer valid. In the
current implementation, we are caching the dialog instance and then
re-using it in media remote view. But if the parent is no longer valid,
it causes a crash when the mDialog.show() is called. Moreover, the
parent activity's instance might be changed in some cases, so, we should
recreate it whenver the show() method in media remote view is called.

Refs: #714